chore: redeem full priced coupons through skill-api routes
