+++ apps/testingaccessibility/.env.development
@@ -1,4 +1,5 @@
 NEXT_PUBLIC_SITE_TITLE=Testing Accessibility
+NEXT_PUBLIC_PRODUCT_NAME=Testing Accessibility
 NEXT_PUBLIC_SELLING_LIVE=true
 NEXTAUTH_URL=http://localhost:3013
 NEXT_PUBLIC_URL=http://localhost:3013
