+++ apps/testingaccessibility/.env.production
@@ -24,6 +24,7 @@ CONVERTKIT_BASE_URL=https://api.convertkit.com/v3/
 NEXT_PUBLIC_GOOGLE_ANALYTICS=G-EZG1S4YEN3
 
 NEXT_PUBLIC_SITE_TITLE=Testing Accessibility
+NEXT_PUBLIC_PRODUCT_NAME=Testing Accessibility
 
 CONVERTKIT_PURCHASED_TAG_ID=3205954
