+++ apps/testingaccessibility/src/server/post-to-slack.ts
@@ -2,7 +2,6 @@ import {WebClient} from '@slack/web-api'
 import {PurchaseInfo} from '../utils/record-new-purchase'
 import {isEmpty} from 'lodash'
 import pluralize from 'pluralize'
-import {getSdk} from '@skillrecordings/database'
 import {Purchase} from '@skillrecordings/database'
 import {postToSlack} from '@skillrecordings/skill-api'
 
@@ -30,32 +29,6 @@ export async function postSaleToSlack(
           }`,
           color: '#eba234',
           title: `Sold (${purchaseInfo.quantity}) ${purchaseInfo.stripeProduct.name}`,
-        },
-      ],
-    })
-  } catch (e) {
-    console.log(e)
-    return false
-  }
-}
-
-export async function postRedemptionToSlack(email: string, productId: string) {
-  const {getProduct} = getSdk()
-  const product = await getProduct({
-    where: {id: productId},
-  })
-
-  try {
-    return await postToSlack({
-      webClient: web,
-      channel: process.env.SLACK_ANNOUNCE_CHANNEL_ID,
-      text: `${email} redeemed ${product?.name}`,
-      attachments: [
-        {
-          fallback: `Redeemed by ${email}`,
-          text: `${email} redeemed a seat!`,
-          color: '#5ceb34',
-          title: `Redeemed ${product?.name}`,
         },
       ],
     })
