+++ packages/skill-api/src/core/types.ts
@@ -9,6 +9,7 @@ export type SkillRecordingsAction =
   | 'prices'
   | 'checkout'
   | 'webhook'
+  | 'redeem'
 
 export interface SkillRecordingsHeader {
   key: string
