+++ packages/skill-api/src/next/index.ts
@@ -102,6 +102,7 @@ export type SlackFeedbackConfig = {
 export type SlackConfig = {
   token: string
   feedback: SlackFeedbackConfig
+  redeem: SlackFeedbackConfig
 }
 
 export interface SkillRecordingsOptions {
